BIIIIIIIG README FIX
Every time I read this repo, I find at least one typo or glaring grammar mistake. I usually ignore them, but I saw that @OtterCodes101 made a PR to fix one singular spelling mistake, and that @doofenator made one somewhat similar to this that got accepted and helped the repo. So (eventually) I made a collection of issues with the README and tried to fix as many as I could find. I'm sure there are still more, but I did all that I could. If you want to go any further, a great strategy would be to entirely delete the Microsoft Labs exploit because wtf is this?


This exploit cos removed in "The Great Reformatting of 2023" so I am gonna readd it!!! Thanks a bunch Bill Gates

Do this! Not drugs!

Go to this website Microsoft Labs
click that button thay say "Sign in to launch VM mode"
Holy guacamole! You need a Microsoft Account! Use your credentials and sign in.
Load the virtual machine by clicking the button or if you sign it it will redirect you to it automatically.
The password is pass@word1
Profit
NOTES: sound doesn't work, a lot of websites are blocked, but its just fun to screw around in it.

(Not really) credit to mundaneunblocking

Even my default Chromebook spellcheck is screaming at me. Anyways, I hope this PR helps the repo, but idk it might get rejected or something I really don't care that much
That's it 👍